requirements.txt
darkdetect==0.8.0
nltk==3.8.1
Pillow==9.2.0
Pillow==10.1.0
PyPDF2==2.10.9
PyPDF2==3.0.1
pytesseract==0.3.10
scikit_learn==1.1.2
tika==2.6.0
typing_extensions==4.8.0
